+++ prototypes/get-number-of-gif-frames.js
@@ -5,8 +5,8 @@
  */
 const im = require('gm').subClass({imageMagick: true});
 
-// const inputPath = `${__dirname}/input/sample.gif`;
-const inputPath = `${__dirname}/input/sample-non-animated.gif`;
+const inputPath = `${__dirname}/input/sample.gif`;
+// const inputPath = `${__dirname}/input/sample-non-animated.gif`;
 
 im(inputPath).identify((err, data) => {
   console.log(JSON.stringify(data, null, 2));